requirements.txt
streamlit
pandas
altair
fpdf
Pillow
altair_saver
openpyxl